Changing float to double

+++ code/logic-cfsd18/sensation/attention/include/attention.hpp
@@ -57,25 +57,25 @@ class Attention : public odcore::base::module::DataTriggeredConferenceClientModu
   void setUp();
   void tearDown();
   
-  void SaveOneCPCPointNoIntensity(const int &pointIndex,const uint16_t &distance_integer, const float &azimuth, const float &verticalAngle, const uint8_t &distanceEncoding);
-  void SaveCPC32NoIntensity(const uint8_t &part, const uint8_t &entriesPerAzimuth, const float &startAzimuth, const float &endAzimuth, const uint8_t &distanceEncoding);
-// void SaveCPC32WithIntensity(const uint8_t &part, const uint8_t &entriesPerAzimuth, const float &startAzimuth, const float &endAzimuth, const uint8_t &distanceEncoding, const uint8_t &numberOfBitsForIntensity, const uint8_t &intensityPlacement, const uint16_t &mask, const float &intensityMaxValue);
+  void SaveOneCPCPointNoIntensity(const int &pointIndex,const uint16_t &distance_integer, const double &azimuth, const double &verticalAngle, const uint8_t &distanceEncoding);
+  void SaveCPC32NoIntensity(const uint8_t &part, const uint8_t &entriesPerAzimuth, const double &startAzimuth, const double &endAzimuth, const uint8_t &distanceEncoding);
+// void SaveCPC32WithIntensity(const uint8_t &part, const uint8_t &entriesPerAzimuth, const double &startAzimuth, const double &endAzimuth, const uint8_t &distanceEncoding, const uint8_t &numberOfBitsForIntensity, const uint8_t &intensityPlacement, const uint16_t &mask, const double &intensityMaxValue);
   void SavePointCloud();
   void ConeDetection();
-  vector<vector<uint32_t>> NNSegmentation(MatrixXf &pointCloudConeROI, const float &connectDistanceThreshold);
-  vector<vector<uint32_t>> FindConesFromObjects(MatrixXf &pointCloudConeROI, vector<vector<uint32_t>> &objectIndexList, const float &minNumOfPointsForCone, const float &maxNumOfPointsForCone, const float &nearConeRadiusThreshold, const float &farConeRadiusThreshold, const float &zRangeThreshold);
-  MatrixXf ExtractConeROI(const float &groundLayerZ, const float &zRangeThreshold, const float &coneHeight);
-  float CalculateXYDistance(MatrixXf &pointCloud, const uint32_t &index1, const uint32_t &index2);
-  float CalculateConeRadius(MatrixXf &potentialConePointCloud);
-  float GetZRange(MatrixXf &potentialConePointCloud);
+  vector<vector<uint32_t>> NNSegmentation(MatrixXf &pointCloudConeROI, const double &connectDistanceThreshold);
+  vector<vector<uint32_t>> FindConesFromObjects(MatrixXf &pointCloudConeROI, vector<vector<uint32_t>> &objectIndexList, const double &minNumOfPointsForCone, const double &maxNumOfPointsForCone, const double &nearConeRadiusThreshold, const double &farConeRadiusThreshold, const double &zRangeThreshold);
+  MatrixXf ExtractConeROI(const double &groundLayerZ, const double &zRangeThreshold, const double &coneHeight);
+  double CalculateXYDistance(MatrixXf &pointCloud, const uint32_t &index1, const uint32_t &index2);
+  double CalculateConeRadius(MatrixXf &potentialConePointCloud);
+  double GetZRange(MatrixXf &potentialConePointCloud);
   void SendingConesPositions(MatrixXf &pointCloudConeROI, vector<vector<uint32_t>> &coneIndexList);
-  opendlv::logic::sensation::Point Cartesian2Spherical(float &x, float &y, float &z);
+  opendlv::logic::sensation::Point Cartesian2Spherical(double &x, double &y, double &z);
   // Define constants to decode CPC message
-  const float START_V_ANGLE = -15.0; //For each azimuth there are 16 points with unique vertical angles from -15 to 15 degrees
-  const float V_INCREMENT = 2.0; //The vertical angle increment for the 16 points with the same azimuth is 2 degrees
-  const float START_V_ANGLE_32 = -30.67; //The starting angle for HDL-32E. Vertical angle ranges from -30.67 to 10.67 degress, with alternating increment 1.33 and 1.34
-  const float V_INCREMENT_32_A = 1.33; //The first vertical angle increment for HDL-32E
-  const float V_INCREMENT_32_B = 1.34; //The second vertical angle increment for HDL-32E
+  const double START_V_ANGLE = -15.0; //For each azimuth there are 16 points with unique vertical angles from -15 to 15 degrees
+  const double V_INCREMENT = 2.0; //The vertical angle increment for the 16 points with the same azimuth is 2 degrees
+  const double START_V_ANGLE_32 = -30.67; //The starting angle for HDL-32E. Vertical angle ranges from -30.67 to 10.67 degress, with alternating increment 1.33 and 1.34
+  const double V_INCREMENT_32_A = 1.33; //The first vertical angle increment for HDL-32E
+  const double V_INCREMENT_32_B = 1.34; //The second vertical angle increment for HDL-32E
 
   // Constants for degree transformation
   const double DEG2RAD = 0.017453292522222; // PI/180.0
@@ -85,9 +85,9 @@ class Attention : public odcore::base::module::DataTriggeredConferenceClientModu
   uint8_t m_12_startingSensorID_32; //From which layer for the first part(12 layers) of CPC for HDL-32E
   uint8_t m_11_startingSensorID_32; //From which layer for the second part(11 layers) of CPC for HDL-32E
   uint8_t m_9_startingSensorID_32; //From which layer for the third part(9 layers) of CPC for HDL-32E
-  array<float, 12>  m_12_verticalAngles; //Store the 12 vertical angles for the first part (including 12 layers) of CPC for HDL-32E
-  array<float, 11> m_11_verticalAngles; //Store the 11 vertical angles for the second part (including 11 layers) of CPC for HDL-32E
-  array<float, 9> m_9_verticalAngles; //Store the 9 vertical angles for the third part (including 9 layers) of CPC for HDL-32E
+  array<double, 12>  m_12_verticalAngles; //Store the 12 vertical angles for the first part (including 12 layers) of CPC for HDL-32E
+  array<double, 11> m_11_verticalAngles; //Store the 11 vertical angles for the second part (including 11 layers) of CPC for HDL-32E
+  array<double, 9> m_9_verticalAngles; //Store the 9 vertical angles for the third part (including 9 layers) of CPC for HDL-32E
   string m_12_cpcDistance_32; //The distance string for the first part of CPC for HDL-32E    
   string m_11_cpcDistance_32; //The distance string for the second part of CPC for HDL-32E    
   string m_9_cpcDistance_32; //The distance string for the third part of CPC for HDL-32E
@@ -107,18 +107,18 @@ class Attention : public odcore::base::module::DataTriggeredConferenceClientModu
   //vector<logic::sensation::Point> pointCloud;
   int m_pointIndex;
   // Define constants and thresolds forclustering algorithm
-  float m_startAngle;
-  float m_endAngle;
-  float m_yBoundary;
-  float m_groundLayerZ;
-  float m_coneHeight;
-  float m_connectDistanceThreshold;
-  float m_layerRangeThreshold;
+  double m_startAngle;
+  double m_endAngle;
+  double m_yBoundary;
+  double m_groundLayerZ;
+  double m_coneHeight;
+  double m_connectDistanceThreshold;
+  double m_layerRangeThreshold;
   uint16_t m_minNumOfPointsForCone;
   uint16_t m_maxNumOfPointsForCone;
-  float m_farConeRadiusThreshold;
-  float m_nearConeRadiusThreshold;
-  float m_zRangeThreshold;
+  double m_farConeRadiusThreshold;
+  double m_nearConeRadiusThreshold;
+  double m_zRangeThreshold;
